rt_manager
Updated max_delay implementation.
Merged manager and manager_rt into rt_manager.
